# docs/schemas/openapi/admin.spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amp Admin API",
...
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Handler for the `GET /jobs` endpoint",
        "description": "Retrieves and returns a paginated list of jobs from the metadata database.\n\n## Query Parameters\n- `limit`: Maximum number of jobs to return (default: 50, max: 1000)\n- `last_job_id`: ID of the last job from previous page for cursor-based pagination\n- `status`: Status filter - \"active\" (default, shows non-terminal jobs), \"all\" (shows all jobs), or comma-separated status values (e.g., \"scheduled,running\")\n\n## Response\n- **200 OK**: Returns paginated job data with next cursor\n- **400 Bad Request**: Invalid limit parameter (0, negative, or > 1000)\n- **500 Internal Server Error**: Database connection or query error\n\n## Error Codes\n- `INVALID_QUERY_PARAMETERS`: Invalid query parameters (malformed or unparseable)\n- `LIMIT_TOO_LARGE`: Limit exceeds maximum allowed value (>1000)\n- `LIMIT_INVALID`: Limit is zero\n- `LIST_JOBS_ERROR`: Failed to list jobs from scheduler (database error)\n- `LIST_JOB_DESCRIPTORS_ERROR`: Failed to list job descriptors from metadata database",
        "operationId": "jobs_list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of jobs to return (default: 50, max: 1000)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "last_job_id",
            "in": "query",
            "description": "ID of the last job from the previous page for pagination",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status filter: 'active' (default, non-terminal jobs), 'all' (all jobs), or comma-separated status values",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Handler for the `POST /jobs` endpoint",
        "description": "Schedules a new job for execution by an available worker. The job type is\ndetermined by the `kind` field in the request body (`gc`, `materialize-raw`,\nor `materialize-derived`). Additional fields depend on the job kind and are\ndefined by the corresponding job descriptor type.\n\n## Request Body\n- `kind`: Job type (`gc`, `materialize-raw`, or `materialize-derived`)\n- `worker_id`: Optional worker selector (exact ID or glob pattern)\n- `retry_strategy`: Optional retry strategy controlling how failed jobs are retried\n- Additional fields vary by job kind (see [`CreateJobRequest`] variants)\n\n## Response\n- **202 Accepted**: Job scheduled successfully\n- **400 Bad Request**: Invalid request body, invalid retry strategy, or no workers\n- **409 Conflict**: An active job with the same idempotency key already exists\n- **500 Internal Server Error**: Scheduler error\n\n## Error Codes\n- `INVALID_BODY`: The request body is not valid JSON, has missing/invalid fields, or has invalid retry strategy parameters\n- `NO_WORKERS_AVAILABLE`: No active workers to handle the job\n- `ACTIVE_JOB_CONFLICT`: An active job with the same idempotency key already exists\n- `SCHEDULER_ERROR`: Internal scheduler failure",
        "operationId": "jobs_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Job scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid body or no workers)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Active job conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Handler for the `DELETE /jobs?status=<filter>` endpoint",
        "description": "Deletes jobs based on status filter. Supports deleting jobs by various status criteria.\n\n## Query Parameters\n- `status=terminal`: Delete all jobs in terminal states (Completed, Stopped, Failed)\n- `status=completed`: Delete all completed jobs\n- `status=stopped`: Delete all stopped jobs\n- `status=error`: Delete all failed jobs\n\n## Response\n- **204 No Content**: Operation completed successfully\n- **400 Bad Request**: Invalid or missing status query parameter\n- **500 Internal Server Error**: Database error occurred\n\n## Error Codes\n- `INVALID_QUERY_PARAM`: Invalid or missing status parameter\n- `DELETE_JOBS_BY_STATUS_ERROR`: Failed to delete jobs by status from scheduler (database error)\n\n## Behavior\nThis handler provides bulk job cleanup with the following characteristics:\n- Only jobs in terminal states (Completed, Stopped, Failed) are deleted\n- Non-terminal jobs are completely protected from deletion\n- Database layer ensures atomic bulk deletion\n- Safe to call even when no terminal jobs exist\n\n## Terminal States\nJobs are deleted when in these states:\n- Completed → Safe to delete\n- Stopped → Safe to delete\n- Failed → Safe to delete\n\nProtected states (never deleted):\n- Scheduled → Job is waiting to run\n- Running → Job is actively executing\n- StopRequested → Job is being stopped\n- Stopping → Job is in process of stopping\n- Unknown → Invalid state\n\n## Usage\nThis endpoint is typically used for:\n- Periodic cleanup of completed jobs\n- Administrative maintenance\n- Freeing up database storage",
        "operationId": "jobs_delete_many",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status filter for jobs to delete",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/String"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Jobs deleted successfully"
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateJobRequest": {
        "oneOf": [
          {
            "allOf": [
              {
                "type": "object",
                "description": "The job descriptor fields (flattened from the GC job descriptor)."
              },
              {
                "type": "object",
                "properties": {
                  "retry_strategy": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "description": "Optional retry strategy controlling how failed jobs are retried."
                  },
                  "worker_id": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "Optional worker selector (exact ID or glob pattern)."
                  }
                }
              },
              {
                "type": "object",
                "description": "Schedule a garbage collection job for a physical table revision.",
                "required": [
                  "kind"
                ],
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "gc"
                    ]
                  }
                }
              }
            ],
            "description": "Schedule a garbage collection job for a physical table revision."
          },
          {
            "allOf": [
              {
                "type": "object",
                "description": "The job descriptor fields (flattened from the raw materialization descriptor)."
              },
              {
                "type": "object",
                "properties": {
                  "retry_strategy": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "description": "Optional retry strategy controlling how failed jobs are retried."
                  },
                  "worker_id": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "Optional worker selector (exact ID or glob pattern)."
                  }
                }
              },
              {
                "type": "object",
                "description": "Schedule a raw dataset materialization job.",
                "required": [
                  "kind"
                ],
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "materialize-raw"
                    ]
                  }
                }
              }
            ],
            "description": "Schedule a raw dataset materialization job."
          },
          {
            "allOf": [
              {
                "type": "object",
                "description": "The job descriptor fields (flattened from the derived materialization descriptor)."
              },
              {
                "type": "object",
                "properties": {
                  "retry_strategy": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "description": "Optional retry strategy controlling how failed jobs are retried."
                  },
                  "worker_id": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "Optional worker selector (exact ID or glob pattern)."
                  }
                }
              },
              {
                "type": "object",
                "description": "Schedule a derived dataset materialization job.",
                "required": [
                  "kind"
                ],
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "materialize-derived"
                    ]
                  }
                }
              }
            ],
            "description": "Schedule a derived dataset materialization job."
          }
        ],
        "description": "Request body for creating a job.\n\nDispatches on the `kind` field to determine the job type.\nFields for each variant are defined by the corresponding job descriptor type."
      },
      "CreateJobResponse": {
        "type": "object",
        "description": "Response body for a created job.",
        "required": [
          "job_id"
        ],
        "properties": {
          "job_id": {
            "type": "integer",
            "format": "int64",
            "description": "The ID of the scheduled job."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by the API\n\nThis struct represents error information returned in HTTP error responses.\nIt provides structured error details including a machine-readable error code\nand human-readable message.\n\n## Error Code Conventions\n- Error codes use SCREAMING_SNAKE_CASE (e.g., `DATASET_NOT_FOUND`)\n- Codes are stable and can be relied upon programmatically\n- Messages may change and should only be used for display/logging\n\n## Example JSON Response\n```json\n{\n  \"error_code\": \"DATASET_NOT_FOUND\",\n  \"error_message\": \"dataset 'eth_mainnet' version '1.0.0' not found\"\n}\n```",
        "required": [
          "error_code",
          "error_message"
        ],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code in SCREAMING_SNAKE_CASE format\n\nError codes are stable across API versions and should be used\nfor programmatic error handling. Examples: `INVALID_SELECTOR`,\n`DATASET_NOT_FOUND`, `METADATA_DB_ERROR`"
          },
          "error_message": {
            "type": "string",
            "description": "Human-readable error message\n\nMessages provide detailed context about the error but may change\nover time. Use `error_code` for programmatic decisions."
          }
        }
      },
...
      "JobInfo": {
        "type": "object",
        "description": "Represents job information for the API response",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "node_id",
          "status",
          "descriptor"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "description": "Job creation timestamp in ISO 8601 / RFC 3339 format"
          },
          "descriptor": {
            "description": "Job descriptor containing job-specific parameters as JSON"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the job (64-bit integer)"
          },
          "node_id": {
            "type": "string",
            "description": "ID of the worker node this job is scheduled for"
          },
          "status": {
            "type": "string",
            "description": "Current status of the job (Scheduled, Running, Completed, Stopped, Failed, etc.)"
          },
          "updated_at": {
            "type": "string",
            "description": "Job last update timestamp in ISO 8601 / RFC 3339 format"
          }
        }
      },
...
      "JobsResponse": {
        "type": "object",
        "description": "API response containing job information",
        "required": [
          "jobs"
        ],
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobInfo"
            },
            "description": "List of jobs"
          },
          "next_cursor": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Cursor for the next page of results (None if no more results)"
          }
        }
      },
...
      "String": {
        "type": "string",
        "description": "Status filter options for job deletion",
        "enum": [
          "Terminal",
          "Completed",
          "Stopped",
          "Error"
        ]
      },
...
    }
  },
...
}